+++ modules/wvp/src/main/java/com/cdzeroly/wvp/gb28181/service/IPlatformChannelService.java
@@ -2,9 +2,10 @@ package com.cdzeroly.wvp.gb28181.service;
 
 import com.cdzeroly.common.mybatis.core.page.PageQuery;
 import com.cdzeroly.common.mybatis.core.page.TableDataInfo;
+import com.cdzeroly.wvp.domain.Group;
 import com.cdzeroly.wvp.gb28181.domian.CommonGBChannel;
 import com.cdzeroly.wvp.gb28181.domian.Platform;
-import com.cdzeroly.wvp.gb28181.domian.Region;
+import com.cdzeroly.wvp.domain.Region;
 import com.cdzeroly.wvp.gb28181.domian.bean.*;
 
 
@@ -16,27 +17,27 @@ import java.util.List;
  */
 public interface IPlatformChannelService {
 
-    TableDataInfo<PlatformChannel> queryChannelList(PageQuery pageQuery, String query, Integer channelType, Boolean online, Integer platformId, Boolean hasShare);
+    TableDataInfo<PlatformChannel> queryChannelList(PageQuery pageQuery, String query, Integer channelType, Boolean online, Long platformId, Boolean hasShare);
 
-    int addAllChannel(Integer platformId);
+    int addAllChannel(Long platformId);
 
-    int removeAllChannel(Integer platformId);
+    int removeAllChannel(Long platformId);
 
-    int addChannels(Integer platformId, List<Integer> channelIds);
+    int addChannels(Long platformId, List<Long> channelIds);
 
-    int removeChannels(Integer platformId, List<Integer> channelIds);
+    int removeChannels(Long platformId, List<Long> channelIds);
 
-    void removeChannels(List<Integer> ids);
+    void removeChannels(List<Long> ids);
 
-    void removeChannel(int gbId);
+    void removeChannel(Long gbId);
 
     List<CommonGBChannel> queryByPlatform(Platform platform);
 
-    void pushChannel(Integer platformId);
+    void pushChannel(Long platformId);
 
-    void addChannelByDevice(Integer platformId, List<Integer> deviceIds);
+    void addChannelByDevice(Long platformId, List<Long> deviceIds);
 
-    void removeChannelByDevice(Integer platformId, List<Integer> deviceIds);
+    void removeChannelByDevice(Long platformId, List<Long> deviceIds);
 
     void updateCustomChannel(PlatformChannel channel);
 
@@ -44,9 +45,9 @@ public interface IPlatformChannelService {
 
     void checkGroupAdd(List<CommonGBChannel> channelList);
 
-    List<Platform> queryPlatFormListByChannelDeviceId(Integer channelId, List<String> platforms);
+    List<Platform> queryPlatFormListByChannelDeviceId(Long channelId, List<String> platforms);
 
-    CommonGBChannel queryChannelByPlatformIdAndChannelId(Integer platformId, Integer channelId);
+    CommonGBChannel queryChannelByPlatformIdAndChannelId(Long platformId, Long channelId);
 
     void checkRegionAdd(List<CommonGBChannel> channelList);
